// tools/yaml-to-toml-converter/client/types.ts
export type YamlToTomlConverterMessages = Readonly<{
  meta: {
    name: string
    description: string
  }
  importFromFileLabel: string
  downloadTomlLabel: string
  yamlLabel: string
  yamlDescription: string
  tomlLabel: string
  tomlDescription: string
  tomlEmptyDescription: string
  yamlPlaceholder: string
  invalidYamlLabel: string
  copyTomlLabel: string
  copiedLabel: string
}>